hello-world-game-py/requirements.txt

viam-sdk==0.55.1

typing-extensions
